Change to saving the admin promote password in redis
For initial creation of an admin account, when no account with the `admin` role
is found on startup, wave creates a password and displays it in the log.

Previously, this value was saved temporarely in a file, /app/admin.txt.
On some deployments e.g. kubernetes this caused crashes, because this location
isnt writable by the "app" user wave is running as.

Now, wave saves this value in redis instead.